Refactor cards
